requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
python-dotenv
